requirements.txt
# Core dependencies
SpeechRecognition==3.10.0
gTTS==2.3.2
pygame==2.5.2
PyAudio==0.2.13

# API integrations
requests==2.31.0
google-generativeai==0.3.1
openai==1.3.5

# Data parsing
beautifulsoup4==4.12.2
pyyaml==6.0.1

# Development tools
python-dotenv==1.0.0
typing-extensions==4.8.0
black==23.11.0
pylint==3.0.2
pytest==7.4.3

# Utilities
pathlib==1.0.1
python-dateutil==2.8.2
tqdm==4.66.1
colorama==0.4.6

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0 